requirements.txt
scc-firewall-manager-sdk
webexpythonsdk
python-dotenv
questionary
rich
requests
pexpect